Update Analysis Desktop Menu
