+++ app/CHANGELOG.md
@@ -3,6 +3,7 @@
 - 2018-03-10 | v.0.17.0 'Phobos':
 
   Update the profile menu on the desktop
+  Update Analysis Desktop Menu
 
 - 2018-03-10 | v.0.16.1 'Phobos':
