+++ app/src/components/DesktopMenu/DesktopAnalysisMenu.css
@@ -2,9 +2,6 @@
   display: flex;
   flex-direction: column;
   white-space: initial;
-  min-width: 600px;
-}
-
-.analysis-menu .analysis-menu-row {
-  display: flex;
+  width: 320px;
+  padding: 12px;
 }
