+++ app/src/components/DesktopMenu/DesktopMenuLinkContainer.css
@@ -27,12 +27,10 @@ a.link-container {
 
 .link-container:hover .link-icon {
   border: 1px solid transparent;
-  border-radius: 2px;
-  box-shadow: 0 7px 9px 0 rgba(76, 95, 114, 0.1);
 }
 
 .link-container .link-title {
-  margin-top: 22px;
+  margin-top: 15px;
   margin-bottom: 6px;
   text-transform: uppercase;
   font-size: 18px;
