+++ app/src/components/TopMenu.js
@@ -3,9 +3,9 @@ import { Link } from 'react-router-dom'
 import 'font-awesome/css/font-awesome.css'
 import { Label } from 'semantic-ui-react'
 import logo from '../assets/logo_sanbase.png'
-import DesktopRightGroupMenu from './DesktopMenu/DesktopRightGroupMenu.js'
 import Search from './Search/SearchContainer'
-import AnalysisDropdownMenu from './AnalysisDropdownMenu'
+import DesktopRightGroupMenu from './DesktopMenu/DesktopRightGroupMenu.js'
+import AnalysisDropdownMenu from './DesktopMenu/AnalysisDropdownMenu'
 import SmoothDropdown from './SmoothDropdown/SmoothDropdown'
 import './AppMenu.css'
 import './TopMenu.css'
